rename: 해당 기능별 type, util 폴더 생성 및 path 적용
