+++ src/utils/register/emailListUtil.ts
